chore(*): Export DisplayDensity common logic to base class (#542)
